rename: 모달컴포넌트 및 에러메시지컴포넌트 폴더구성 변경
